+++ src/actions/SaveSubmission.js
@@ -2,9 +2,8 @@
 
 const _ = require('lodash');
 const async = require('async');
-const {VM} = require('vm2');
 const util = require('../util/util');
-
+const {evaluateSync} = require('@formio/vm');
 const LOG_EVENT = 'Save Submission Action';
 
 module.exports = function(router) {
@@ -204,19 +203,15 @@ module.exports = function(router) {
 
         if (this.settings.transform) {
           try {
-            let vm = new VM({
-              timeout: 500,
-              sandbox: {
+            const newData = evaluateSync({
+              deps: [],
+              code: this.settings.transform,
+              data: {
                 submission: (res.resource && res.resource.item) ? res.resource.item : req.body,
                 data: submission.data,
-              },
-              eval: false,
-              fixAsync: true
+              }
             });
-
-            const newData = vm.run(this.settings.transform);
             submission.data = newData;
-            vm = null;
           }
           catch (err) {
             debug(`Error in submission transform: ${err.message}`);